+++ react.js
@@ -2,21 +2,7 @@
 // add further configs
 
 module.exports = {
-  "settings": {
-    "react": {
-      "version": "detect", // To remove the warning.
-    }
-  },
-  "plugins": [
-    "react",
-  ],
   "extends": [
-    "plugin:react/recommended",
-    "plugin:react-hooks/recommended",
-    "./ts.js",
+    "./react-common.js"
   ],
-  "rules": {
-    "react/prop-types": "off",
-    "react/display-name": "off",
-  }
 };